Conflicts:
	Distribution.xcodeproj/project.pbxproj
	Distribution/MDViewController/PreparePay/MDPreparePayViewController.m

+++ Distribution/MDViewController/RequestDetail/MDRequestDetailViewController.m
@@ -27,7 +27,7 @@
 
 -(void)initNavigationBar {
     NSString *number = [NSString stringWithFormat:@"%@",_data[@"package_number"]];
-    int length = number.length/2;
+    int length = (int)number.length/2;
     NSString *numberLeft = [number substringToIndex:length];
     NSString *numberRight = [number substringFromIndex:length];
     self.navigationItem.title = [NSString stringWithFormat:@"番号: %@ - %@",numberLeft, numberRight];
